+++ base/repl/REPLCompletions.jl
@@ -105,7 +105,7 @@ const sorted_keywords = [
     "true", "try", "using", "while"]
 
 function complete_keyword(s::Union{String,SubString{String}})
-    r = searchsorted(sorted_keywords, s)
+    r = findsorted(s, sorted_keywords)
     i = first(r)
     n = length(sorted_keywords)
     while i <= n && startswith(sorted_keywords[i],s)